Hey, do you want a riddle? What will be this commit? :dancer: More things about the powerups
Solved a bug concerning hitboxes. Player was able to go through some of
them because of the Collider::ColliderHit function.

Added the loot of Extended Energy bonus + unloaded sounds in
ModuleEnemies.cpp

Adjusted the camera so it will move with the actual speed of the player
in ModulePlayer.cpp

Energy bar now swaps each time you pick a extended bar bonus / you die.